requirements.txt
psycopg2-binary==2.9.3
Django==4.0.3
python-decouple==3.6
gunicorn==20.1.0
gevent==21.12.0
django-more-admin-filters==1.3
django-nested-admin==3.4.0